fix: keep functional coverage tests within repository boundaries
